Programmed with the wireless debug demo
This is the same program I ran last night. Use the same .ioc file and everything should work out.
During operation, pressing the user button on one board will turn on the green LED of the other board wirelessly, assuming the configuration is correct.
Note that this requires the SCK connection in PB_12, NOT PA_12, as shown currently in the config diagram. I'll update that later today.